+++ src/layout/MainLayout/MainLayout.tsx
@@ -2,12 +2,12 @@ import React, { useEffect, useState, ReactNode } from 'react';
 import { Chat, SidebarMenu, Header } from '../../components';
 import style from './MainLayout.module.css';
 import { getIsSidebarOpen, setIsSidebarOpen } from '../../utils';
+import Settings from '../../components/Settings/Settings';
 
 interface IMainLayout {
-  onOpenSettings: () => void;
   children: ReactNode;
 }
-const MainLayout = ({ onOpenSettings, children }: IMainLayout) => {
+const MainLayout = ({ children }: IMainLayout) => {
   const mediaQuery = window.matchMedia('(max-width: 768px)');
   const initIsSidebarVisible = mediaQuery.matches
     ? !mediaQuery.matches
@@ -17,6 +17,11 @@ const MainLayout = ({ onOpenSettings, children }: IMainLayout) => {
   const [isSidebarVisible, setIsSidebarVisible] =
     useState(initIsSidebarVisible);
   const [isOverlayed, setIsOverlayed] = useState(mediaQuery.matches);
+  const [isSettingsOpen, setIsSettingsOpen] = useState(false);
+
+  const toggleSettingsModal = () => {
+    setIsSettingsOpen((prev) => !prev);
+  };
 
   const toggleSidebar = () => {
     setIsSidebarVisible(!isSidebarVisible);
@@ -47,25 +52,31 @@ const MainLayout = ({ onOpenSettings, children }: IMainLayout) => {
   }, []);
 
   return (
-    <div className={style.container}>
-      <SidebarMenu
-        isSidebarVisible={isSidebarVisible}
-        onClose={toggleSidebar}
-        onOpenSettings={onOpenSettings}
-      />
-      <Header
-        isSidebarVisible={isSidebarVisible}
-        onOpenSidebar={toggleSidebar}
-      />
-      <div
-        className={`${style.childrenContainer} ${isSidebarVisible ? style.sidebarVisible : style.sidebarHidden}`}
-      >
-        {children}
+    <>
+      <div className={style.container}>
+        <SidebarMenu
+          isSidebarVisible={isSidebarVisible}
+          onClose={toggleSidebar}
+          onOpenSettings={toggleSettingsModal}
+        />
+        <Header
+          isSidebarVisible={isSidebarVisible}
+          onOpenSidebar={toggleSidebar}
+        />
+        <div
+          className={`${style.childrenContainer} ${isSidebarVisible ? style.sidebarVisible : style.sidebarHidden}`}
+        >
+          {children}
+        </div>
+        {isSidebarVisible && isOverlayed && (
+          <div className={style.overlay} onClick={toggleSidebar}></div>
+        )}
       </div>
-      {isSidebarVisible && isOverlayed && (
-        <div className={style.overlay} onClick={toggleSidebar}></div>
-      )}
-    </div>
+      <Settings
+        openSettings={isSettingsOpen}
+        closeSettings={toggleSettingsModal}
+      />
+    </>
   );
 };
