+++ src/pages/MainPage/MainPage.tsx
@@ -1,16 +1,12 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { Chat } from '../../components';
 import { useThreads } from '../../hooks';
 import { getCurrentThreadId } from '../../utils';
 import { MainLayout } from '../../layout';
-import Settings from '../../components/Settings/Settings';
 
 const MainPage = () => {
   const { initChatting, openThread } = useThreads();
-  const [isSettingsOpen, setIsSettingsOpen] = useState(false);
-  const toggleSettingsModal = () => {
-    setIsSettingsOpen((prev) => !prev);
-  };
+
   useEffect(() => {
     const currentThreadId = getCurrentThreadId();
     currentThreadId === -1 || currentThreadId === null
@@ -20,13 +16,9 @@ const MainPage = () => {
 
   return (
     <>
-      <MainLayout onOpenSettings={toggleSettingsModal}>
+      <MainLayout>
         <Chat />
       </MainLayout>
-      <Settings
-        openSettings={isSettingsOpen}
-        closeSettings={toggleSettingsModal}
-      />
     </>
   );
 };
